Refactor(organisation): move functions and dependencies in separated files
